% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data2.R
\name{actg_tsv_to_gdata}
\alias{actg_tsv_to_gdata}
\title{actg_tsv_to_gdata}
\usage{
actg_tsv_to_gdata(geno_path, scans_path, scans_col_map = "scan_id", snps_path,
  snps_col_map = c("probe_id", "chromosome", "position"),
  na_encoding = "--")
}
\arguments{
\item{geno_path}{Path of genotype text tsv file.
First line is scan ids and first column is snps ids.
NAs are specified via na_encoding parameter.
Header is managed automatically by data.table::fread.}

\item{scans_path}{Path of scans text tsv file.
Header is managed automatically by data.table::fread.}

\item{scans_col_map}{Character vector mapping the columns of the scans file,
the index in the vector is the column index in the text
file, the characters are the names for the gds file
NAs specify undesirable columns.
  Required characters: 'scan_id'}

\item{snps_path}{Path of snps text tsv file
Header is managed automatically by data.table::fread.}

\item{snps_col_map}{Character vector mapping the columns of the snpss file,
the index in the vector is the column index in the text
file, the characters are the names for the gds file.
NAs specify undesirable columns.
  Required characters: probe_id, chromosome, position}

\item{na_encoding}{Vector of characters encoding NAs in genotype file,}
}
\value{
GenotypeData object
}
\description{
Converts TSV files (geno, SNPs, scans) with genotypes as ACTG characters
to GWASTools GenotypeData.
Jun 4, 2014
}
\author{
tcharlon
}

