% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{add_on_exit}
\alias{add_on_exit}
\title{set an on.exit handler in another function/env}
\usage{
add_on_exit(expr, where = parent.frame())
}
\arguments{
\item{expr}{the expression to be executed by on.exit. the expression
must be evaluable in the where environment}

\item{where}{in which function environment to execute/set the on.exit}
}
\description{
on.exit can normally only be set in the current function.
Trying to set it using eval() or delayed evaluation does not work
}
\details{
See this dicussion:
\url{https://stat.ethz.ch/pipermail/r-devel/2013-November/067874.html}
}
\seealso{
\code{\link{setup_temp_dir}}
}

