% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{add_pca_sup_vars}
\alias{add_pca_sup_vars}
\title{Add supplementary vars to an existing qb_pca object.}
\usage{
add_pca_sup_vars(qb_pca, df_vars)
}
\arguments{
\item{qb_pca}{qb_pca object}

\item{df_vars}{variables data frame}
}
\value{
qb_pca object with additional supplementary variables
}
\description{
The new variables need to be supplied as a data.frame
containing an identifier column with the same name as
the one contained in the qb_pca object.
The data is added using a left join (old-data, new-data).
This ensures that the original data is not subsetted.
}
\author{
adicara
}

