% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data2.R
\name{bed_targz_to_gds}
\alias{bed_targz_to_gds}
\title{bed_targz_to_gds}
\usage{
bed_targz_to_gds(tarpath, gdspath)
}
\arguments{
\item{tarpath}{path of input tar file}

\item{gdspath}{path of output gds file}
}
\description{
Untar and call snpgdsBED2GDS
Sep 24, 2014
}
\author{
tcharlon
}

