% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data2.R
\name{build_gwastools}
\alias{build_gwastools}
\title{build_gwastools}
\usage{
build_gwastools(geno, scans, snps)
}
\arguments{
\item{geno}{genotype data frame}

\item{scans}{scans data frame}

\item{snps}{snps data frame}
}
\value{
genotype data object
}
\description{
wrapper for GWASTools functions to build GenotypeData object
Aug 2, 2014
}
\author{
tcharlon
}

