% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{catch_warnings}
\alias{catch_warnings}
\title{suppress the warnings, and return them along with the result of the expression}
\usage{
catch_warnings(expr)
}
\arguments{
\item{expr}{expression to evaluate.}
}
\value{
a list with 2 elements: the result of the evaluated expression,
   and the list of caught warning messages
}
\description{
suppress the warnings, and return them along with the result of the expression
}
\seealso{
Other exceptions: \code{\link{die_if}};
  \code{\link{die_unless}}
}

