% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arg}
\alias{check_arg}
\title{check that a value is conform to our argument format}
\usage{
check_arg(spec, value, argname = deparse(substitute(value)), call = "")
}
\arguments{
\item{spec}{the check specification string}

\item{value}{the value to check against the format}

\item{argname}{the name of the argument to check}

\item{call}{the function call, for error message clarity}
}
\value{
TRUE
}
\description{
die if the argument is invalid
}
\details{
The format string is made of
 \describe{
     \item{not_null:optional}{ a ! means that NULL is not allowed }

\item{type:mandatory}{ a one letter type descriptor. upper-case means
         that NAs are not allowed.
      \describe{
         \item{i}{integer}
         \item{n}{numeric}
         \item{c}{character}
         \item{s}{string: non empty strings}
         \item{b}{logical (boolean) }
         \item{l}{list}
         \item{d}{data.frame}
         \item{a}{any type}
       }}

\item{length:mandatory}{
         N.B:For a data.frame, the length is the number of columns
         does not apply to NULL value, use ! to forbid NULL
      \describe{
         \item{*}{ any length }
         \item{+}{ length > 0 }
         \item{?}{ 0 or 1 }
         \item{number+}{ length equal or higher than the indicated number}
         \item{number expression}{ a R expression that must evaluate to
             something convertible to integer }
       }}

}
}
\examples{
check_arg('N+', 1:5)
check_arg('!S+', "coucou")
check_arg('i1:10', 1:5)
check_arg('i2+', 1:5)
}
\seealso{
Other args_checking: \code{\link{check_fx_args}}
}

