% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{check_availability}
\alias{check_availability}
\title{check availability}
\usage{
check_availability(selected, available, name_error_msg)
}
\arguments{
\item{selected}{logical or character vector, see decription}

\item{available}{logical or character vector, see decription}

\item{name_error_msg}{name of the character vector printed in the error message}
}
\value{
selected character vector, see description
}
\description{
check that elements of selected vector are in the available vector,
or if selected is true get all available, if false get none.
}
\details{
Jun 5, 2014
}
\author{
tcharlon
}

