% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{check_fx_args}
\alias{check_fx_args}
\title{check the function arguments using a concise syntax}
\usage{
check_fx_args(...)
}
\arguments{
\item{...}{specifications for arguments}
}
\description{
Use the \code{\link{check_arg}} syntax for each checked argument
}
\examples{
f <- function(x, y = 1, z = NULL) {
   check_fx_args(x = "I+", y = "n1", z = "C*")
}
}
\seealso{
Other args_checking: \code{\link{check_arg}}
}

