% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{check_genotype_data}
\alias{check_genotype_data}
\title{check a GenotypeData object}
\usage{
check_genotype_data(gdata)
}
\arguments{
\item{gdata}{a GenotypeData object}
}
\description{
die in case of problem
}
\author{
Karl
}

