% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{check_snp_gds}
\alias{check_snp_gds}
\title{check a GenotypeData object}
\usage{
check_snp_gds(gds)
}
\arguments{
\item{gds}{a gds filename or object}
}
\description{
die in case of problem
}
\author{
Karl
}

