% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{compute_pca_coords_vars_sup}
\alias{compute_pca_coords_vars_sup}
\title{Get the coordinates for supplementary variables}
\usage{
compute_pca_coords_vars_sup(qb_pca, vars)
}
\arguments{
\item{qb_pca}{qb_pca object}

\item{vars}{vector of supplementary variable name for which the
coordinates need to be computed. By default all the supplementary
variables contained in the qb_pca object.}
}
\value{
data.frame containing the PC values for each supplementary variable.
Categorical variables are represented by level
}
\description{
Get the coordinates for supplementary variables
}
\author{
adicara
}

