% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{compute_pca_obs_sup}
\alias{compute_pca_obs_sup}
\title{Compute the PC coordinates of supplementary observations}
\usage{
compute_pca_obs_sup(qb_pca_obj, obs)
}
\arguments{
\item{qb_pca_obj}{qb_pca object}

\item{obs}{character vector of obs to retrieve}
}
\value{
data.frame with PC coordinates of sup individuals
}
\description{
Compute the PC coordinates of supplementary observations
}
\author{
adicara
}

