% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{compute_pca_prcomp}
\alias{compute_pca_prcomp}
\title{Compute the pca using prcomp}
\usage{
compute_pca_prcomp(data, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{data.frame in wide format}

\item{center}{Parameter passed to prcomp as center.}

\item{scale}{Parameter passed to prcomp as scale..}
}
\value{
prcomp object
}
\description{
Compute the pca using prcomp
}
\author{
adicara
}

