% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{create_pca_axis_labels}
\alias{create_pca_axis_labels}
\title{Create labels for PCA axes, adding if required pct of explained var}
\usage{
create_pca_axis_labels(qb_pcafort, pc_variance = TRUE, axes,
  names_axe = paste0("PC", axes))
}
\arguments{
\item{qb_pcafort}{qb_pcafort or qb_pca object (which is fortified)}

\item{pc_variance}{Logical, displays the variance ratio along the axis names}

\item{axes}{Two principal component axes to plot.
Default: 1:2}

\item{names_axe}{names of PCA axes}
}
\value{
character vector of axis labels
}
\description{
Create labels for PCA axes, adding if required pct of explained var
}
\author{
tcharlon
}

