% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{df_columns_exist}
\alias{df_columns_exist}
\title{Checks if one or more column names exist in the data.frame}
\usage{
df_columns_exist(data_frame, cols, silent = FALSE)
}
\arguments{
\item{data_frame}{the data frame to check columns in}

\item{cols}{non empty vector of column names}

\item{silent}{display warning (default "TRUE")}
}
\value{
boolean
}
\description{
TRUE iff column found + warning with name(s) of missing column(s)
}

