% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc_pca.R
\name{df_rbind_all}
\alias{df_rbind_all}
\title{rbind data frames (fast)}
\usage{
df_rbind_all(..., use_row_names = FALSE)
}
\arguments{
\item{...}{Data frames to combine.

  You can either supply one data frame per argument, or a list of
  data frames in the first argument.

  When column-binding, rows are matched by position, not value so all data
  frames must have the same number of rows. To match by value, not
  position, see \code{left_join} etc. When row-binding, columns are
  matched by name, and any values that don't match will be filled with NA.}

\item{use_row_names}{whether to use the row names, otherwise
the rbound data frame will have default rownames}
}
\value{
a data frame
}
\description{
currently a wrapper around dplyr::bind_rows
}
\details{
Beware: the resulting columns will be factor if and only if the
levels are identical
}

