% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{die_if}
\alias{die_if}
\title{die (stop) if the condition is TRUE, with a \code{\link{sprintf}}
formatted message}
\usage{
die_if(cond, format, ...)
}
\arguments{
\item{cond}{the condition that should be met}

\item{format}{the formatted message}

\item{...}{additional args for \code{\link{sprintf}}}
}
\description{
die (stop) if the condition is TRUE, with a \code{\link{sprintf}}
formatted message
}
\seealso{
Other exceptions: \code{\link{catch_warnings}};
  \code{\link{die_unless}}
}

