% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{die_unless}
\alias{die_unless}
\title{die (stop) unless the condition is TRUE, with a \code{\link{sprintf}}
formatted message}
\usage{
die_unless(cond, format, ...)
}
\arguments{
\item{cond}{the condition that should be met}

\item{format}{the formatted message}

\item{...}{additional arguments for \code{\link{sprintf}}.
               If some arguments have length > 1, they are automatically
               collapsed into a string}
}
\description{
die (stop) unless the condition is TRUE, with a \code{\link{sprintf}}
formatted message
}
\examples{
\dontrun{
 s <- c('A', 'b', 'C', 'd')
 res <- try(die_unless(s \%in\% LETTERS, 'bad letters: "\%s"',
      s[!s\%in\% LETTERS]), silent = TRUE)
 # "Error : bad letters: \\"b,d\\"\\n"


}
}
\seealso{
Other exceptions: \code{\link{catch_warnings}};
  \code{\link{die_if}}
}

