% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{draw_pca_vars}
\alias{draw_pca_vars}
\title{Draw variables on PC space}
\usage{
draw_pca_vars(data_vars, names_axe, scale = 1, circle = FALSE,
  scale_sdev = FALSE)
}
\arguments{
\item{data_vars}{qb_pca object subsetted on PCA_VARTYPE == 'VAR' or 'VAR_SUP'}

\item{names_axe}{name of columns{2} containing PCs to plot}

\item{scale}{scale of the values}

\item{circle}{whether to draw a surrounding circle}

\item{scale_sdev}{Should the values be scaled by the std. dev. of PCs ?}
}
\value{
list of geoms
}
\description{
Draw variables on PC space
}
\author{
adicara
}

