% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_gds.default}
\alias{fetch_gds.default}
\title{default method}
\usage{
\method{fetch_gds}{default}(obj, ...)
}
\arguments{
\item{obj}{an object containing genotypes}

\item{...}{additional arguments}
}
\description{
default method
}

