% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_gds.GdsGenotypeReader}
\alias{fetch_gds.GdsGenotypeReader}
\title{for GdsGenotypeReader}
\usage{
\method{fetch_gds}{GdsGenotypeReader}(obj, ...)
}
\arguments{
\item{obj}{an object containing genotypes}

\item{...}{additional arguments}
}
\description{
for GdsGenotypeReader
}

