% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_gds.GenotypeData}
\alias{fetch_gds.GenotypeData}
\title{for GenotypeData}
\usage{
\method{fetch_gds}{GenotypeData}(obj, ...)
}
\arguments{
\item{obj}{an object containing genotypes}

\item{...}{additional arguments}
}
\description{
for GenotypeData
}

