% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_gds.GenotypeDataSubset}
\alias{fetch_gds.GenotypeDataSubset}
\title{for GenotypeDataSubset: no directly connected GDS}
\usage{
\method{fetch_gds}{GenotypeDataSubset}(obj, ...)
}
\arguments{
\item{obj}{an object containing genotypes}

\item{...}{additional arguments}
}
\description{
for GenotypeDataSubset: no directly connected GDS
}

