% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{fetch_genotypes}
\alias{fetch_genotypes}
\title{Replacement for GenotypeData::getGenotype}
\usage{
fetch_genotypes(gdata, snps_idx = NULL, scans_idx = NULL, char = FALSE,
  snps_first = is_snp_first_dim(gdata))
}
\arguments{
\item{gdata}{a GenotypeData object}

\item{snps_idx}{vector of snps indices}

\item{scans_idx}{vector of scans indices}

\item{char}{to retrieve genotypes as characters using stored alleles
                   if any or default "A" and "B"}

\item{snps_first}{whether snps is the first dimension. to optimize
the fetch strategy}
}
\value{
genotypes as a matrix snps by scans in the same order as snp_idx
 and scans_idx
}
\description{
Easier: use indices rather that intervals for snps
Fast:   indices are split in intervals for fact chunk reads
Safer:  never drop dimensions
Better: able to get genotype strings using alleles from GDS-based
   GenotypeData
}
\details{
will die if no snps or no scans are selected
}
\author{
karl (based on adicara fetch_genotypes_by_block)
}

