% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{fetch_genotypes_by_blocks}
\alias{fetch_genotypes_by_blocks}
\title{Retrieve genotypes from a GenotypeData object using consecutive blocks}
\usage{
fetch_genotypes_by_blocks(gdata, snps_blocks = list(c(1, nsnp(gdata))),
  scans_blocks = list(c(1, nscan(gdata))),
  snps_first = is_snp_first_dim(gdata))
}
\arguments{
\item{gdata}{a GenotypeData object}

\item{snps_blocks}{a list of disjoint blocks (i.e c(start, length) of snps}

\item{scans_blocks}{a list of disjoint blocks (i.e c(start, length) of scans}

\item{snps_first}{whether snps is the first dimension. to optimize
the fetch strategy}
}
\value{
see \code{\link{fetch_genotypes}}
}
\description{
Retrieve genotypes from a GenotypeData object using consecutive blocks
}
\author{
karl
}

