% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{fetch_genotypes_by_chunk_iterator}
\alias{fetch_genotypes_by_chunk_iterator}
\title{return an iterator that fetches the genotypes by chunks}
\usage{
fetch_genotypes_by_chunk_iterator(gdata, chunk_size = 100L,
  nb_chunks = NULL)
}
\arguments{
\item{gdata}{a GenotypeData object}

\item{chunk_size}{the desired chunk size}

\item{nb_chunks}{the desired nb_chunks. Takes precedence over chunk_size
 if both given}
}
\value{
an iterator as a list with fields nb, chunk_size, get(i)
}
\description{
return an iterator that fetches the genotypes by chunks
}
\author{
karl
}

