% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{fetch_genotypes_by_idx}
\alias{fetch_genotypes_by_idx}
\title{Retrieve genotypes from a GenotypeData object using snp and scan indices}
\usage{
fetch_genotypes_by_idx(gdata, snps_idx = seq_len(nsnp(gdata)),
  scans_idx = seq_len(nscan(gdata)), snps_first = is_snp_first_dim(gdata))
}
\arguments{
\item{gdata}{a GenotypeData object}

\item{snps_idx}{vector of snps indices}

\item{scans_idx}{vector of scans indices}

\item{snps_first}{whether snps is the first dimension. to optimize
the fetch strategy}
}
\value{
see \code{\link{fetch_genotypes}}
}
\description{
Perform smart queries by reordering the indices into consecutive blocks
}
\author{
karl
}

