% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{gds_genotype_reader}
\alias{gds_genotype_reader}
\title{alternative constructor for GdsGenotypeReader that also accepts a gds object}
\usage{
gds_genotype_reader(gds, ...)
}
\arguments{
\item{gds}{a gds filename or object}

\item{...}{additional args for GdsGenotypeReader constructor}
}
\description{
This hopefully will be implemented in GdsGenotypeReader if my patch is accepted
When given a filename, will open it with allow_duplicate and allow_fork.
You will then be responsible for closing the handle, but because it's opened
with allow_duplicate it is not that important
}
\author{
Karl
}

