% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2_all_generics.R
\docType{methods}
\name{All generics}
\alias{All generics}
\alias{fetch_allele1}
\alias{fetch_allele2}
\alias{fetch_gds}
\alias{is_snp_first_dim}
\title{test if the genotypes first dimension is snp}
\usage{
is_snp_first_dim(obj, ...)

fetch_gds(obj, ...)

fetch_allele1(obj, snps_idx)

fetch_allele2(obj, snps_idx)
}
\arguments{
\item{obj}{an object containing genotypes}

\item{...}{additional arguments}

\item{snps_idx}{the indices of the snps for which to fetch}
}
\value{
TRUE iff snp is the first dimension

a one-character vector of alleles, or NULL if no alleles

a one-character vector of alleles
}
\description{
The genotypes matrix can either be snp*sample or sample*snp
It is more efficient is the matrix is to be accessed by snp to
store it  sample*snp, i.e. is_gds_snp_first_dim == FALSE

get the GDS object associated with a GenotypeData object if any

fetch the first alleles

fetch the second alleles
}

