% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{genotype_data_subset}
\alias{genotype_data_subset}
\title{alternative constructor for GenotypeDataSubsetSubset}
\usage{
genotype_data_subset(gdata, snps_idx = seq_len(nsnp(gdata)),
  scans_idx = seq_len(nscan(gdata)))
}
\arguments{
\item{gdata}{a GenotypeData object}

\item{snps_idx}{indices of the SNPs to keep}

\item{scans_idx}{indices of the scans to keep}
}
\value{
a GenotypeDataSubset object
}
\description{
alternative constructor for GenotypeDataSubsetSubset
}
\author{
Karl
}

