% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{get_pca}
\alias{get_pca}
\title{Get qb_pca object}
\usage{
get_pca(data, id_col, obs = NULL, vars = NULL, center = TRUE,
  scale = TRUE)
}
\arguments{
\item{data}{Data frame subsetted in columns by vars, in rows by obs,
and passed as parameter data to prcomp.}

\item{id_col}{Name of identifier column for observations (must not contain duplicates)}

\item{obs}{Row subset of data. No subset if NULL.
Default: NULL}

\item{vars}{Names of columns to be used for the PCA. All the variables
which are not specified will be considered as supplementary variables. If NULL
all the variables are used except the ID column}

\item{center}{Parameter passed to prcomp as center.}

\item{scale}{Parameter passed to prcomp as scale..}
}
\value{
List of data, variables, and prcomp object.
}
\description{
Create a qb_pca object that stores the pca object and the
original data
}

