% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{get_pca_explained_var}
\alias{get_pca_explained_var}
\title{Get explained variance for each PC}
\usage{
get_pca_explained_var(qb_pca)
}
\arguments{
\item{qb_pca}{qb_pca object}
}
\value{
vector of explained variance for each PC expressed in percentage
}
\description{
Get explained variance for each PC
}
\author{
adicara
}

