% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{get_pca_individuals}
\alias{get_pca_individuals}
\title{Extract individuals from the PCA, together
with the PCs values}
\usage{
get_pca_individuals(qb_pca, active = TRUE, sup = TRUE)
}
\arguments{
\item{qb_pca}{qb_pca object}

\item{active}{Logical or vector of active observations names to retrieve.
If true all are retrieved, if false none}

\item{sup}{Logical or vector of supplementary observations names to retrieve.
If true all are retrieved, if false none}
}
\value{
data.frame of individuals
}
\description{
Extract individuals from the PCA, together
with the PCs values
}
\author{
adicara
}

