% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{get_pca_panels}
\alias{get_pca_panels}
\title{Get PCA pairs panels}
\usage{
get_pca_panels(pca, seq_axes = 1:4, groups = NULL, max_vars = 0L, ...)
}
\arguments{
\item{pca}{pca or qb_pca object (which is fortified)}

\item{seq_axes}{Sequence of axes to display}

\item{groups}{Column index or name of qb_pca$data for the grouping of
observations.
  Default: NULL}

\item{max_vars}{Number of variables to display}

\item{...}{Passed to ggplot_pca}
}
\description{
Get the pair combinations panels of Principal Components.
}

