% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2_all_generics.R
\name{get_scan_annot}
\alias{get_scan_annot}
\title{accessor for the scanAnnot slot of GenotypeData}
\usage{
get_scan_annot(obj, ...)
}
\arguments{
\item{obj}{an GenotypeData like object}

\item{...}{additional arguments}
}
\value{
a ScanAnnotationDataFrame object, or NULL
}
\description{
accessor for the scanAnnot slot of GenotypeData
}

