% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/2_all_generics.R
\name{get_snp_annot}
\alias{get_snp_annot}
\title{accessor for the snpAnnot slot of GenotypeData}
\usage{
get_snp_annot(obj, ...)
}
\arguments{
\item{obj}{an GenotypeData like object}

\item{...}{additional arguments}
}
\value{
a SnpAnnotationDataFrame object, or NULL
}
\description{
accessor for the snpAnnot slot of GenotypeData
}

