% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplot_manhat}
\alias{ggplot_manhat}
\title{ggplot_manhat}
\usage{
ggplot_manhat(df_vars, gdata, axes = 1:9, byposition = TRUE,
  n_points = 2000, ...)
}
\arguments{
\item{df_vars}{Data frame of variable contributions with columns
PCA_VARNAME and PC{axes}}

\item{gdata}{GenotypeData object with snpIDs matching df_vars column
PCA_VARNAME}

\item{axes}{Axes to plot}

\item{byposition}{Plot the SNPs by chromosome and poistion or by index}

\item{n_points}{Maximum number of largest values to plot, for each
principal component}

\item{...}{Passed to ggplot2::facet_wrap}
}
\value{
ggplot
}
\description{
ggplot_manhat
}

