% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplot_pca}
\alias{ggplot_pca}
\title{Ggplot a qb_pca object}
\usage{
ggplot_pca(pca, groups = NULL, axes = c(1, 2), obs = TRUE,
  obs_sup = FALSE, vars = FALSE, vars_sup = FALSE, pc_variance = TRUE,
  ellipses = FALSE, label = NULL, link = FALSE, alpha_fill = FALSE,
  ellipses_ci = 0.95, white_panel = TRUE, nas_first = TRUE, ...)
}
\arguments{
\item{pca}{pca or qb_pca object (which is fortified)}

\item{groups}{Column index or name of qb_pca$data for the grouping of
observations.
  Default: NULL}

\item{axes}{Two principal component axes to plot.
Default: 1:2}

\item{obs}{Character identifiers of active observations to use. If TRUE no
subset. If FALSE none are used.
  Default: TRUE}

\item{obs_sup}{Character identifiers of suppl. observations to use. If TRUE no
subset. If FALSE none are used.
  Default: FALSE}

\item{vars}{Character identifiers of active variables to use. If TRUE no
subset. If FALSE none are used.
  Default: FALSE}

\item{vars_sup}{Character identifiers of suppl. variables to use. If TRUE no
subset. If FALSE none are used.
  Default: FALSE}

\item{pc_variance}{Logical, displays the variance ratio along the axis names}

\item{ellipses}{Should ellipses be plotted ? Not used if groups is NULL.
Default: FALSE}

\item{label}{name of column to be used for labelling observation. Can be
a supplementary variable or "PCx". By default (NULL), no labels are shown.}

\item{link}{if TRUE a line is drawn between the observations of the same
group. This is useful to identify linked observations}

\item{alpha_fill}{Logical, should also use alpha to discriminate active and
additional observations ?
  Default: FALSE}

\item{ellipses_ci}{Confidence interval ratio of the ellipses. Not used if
ellipses is FALSE.
  Default: .95}

\item{white_panel}{Logical, changes standard panel theme}

\item{nas_first}{Logical, rearranges data frame to put NAs in groups first}

\item{...}{Passed to draw_pca_vars}
}
\value{
ggplot object
}
\description{
Ggplot a qb_pca object
}
\author{
tcharlon
}

