% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{grob_pca_panels}
\alias{grob_pca_panels}
\title{Plot combinations of PCs, with variance contributions}
\usage{
grob_pca_panels(plots, only_panels = TRUE, legend = FALSE)
}
\arguments{
\item{plots}{List of ggplots}

\item{only_panels}{Logical, returns only the plot panels}

\item{legend}{Should the legend be kept ?}
}
\value{
Gtable
}
\description{
Jun 22, 2014
}
\author{
tcharlon
}

