% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dim_red.R
\name{haplo_em}
\alias{haplo_em}
\title{haplo_em}
\usage{
haplo_em(gdata, snp_idxs, impute = sample_impute,
  em_control = haplo.stats::haplo.em.control(min.posterior = 0.001, n.try =
  1))
}
\arguments{
\item{gdata}{GenotypeData object}

\item{snp_idxs}{SNPs indexes}

\item{impute}{Function for imputation. If NULL no imputation.}

\item{em_control}{haplo.stats::haplo.em.control return}
}
\value{
Numeric matrix (haplotypes x frequency,rs_ids,hap{1,2}code.scan_id)
and an optional pvalues column (depending on pvals parameter).
}
\description{
Get haplotypes with iterative EM from haplo.stats (quadratic complexity).
Jan 5, 2015
}
\author{
tcharlon
}

