% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dim_red.R
\name{haplo_mcmc}
\alias{haplo_mcmc}
\title{haplo_mcmc}
\usage{
haplo_mcmc(path, probe_ids, n_cores = 1)
}
\arguments{
\item{path}{BED (optionally tar gz) file path}

\item{probe_ids}{Probe ids of SNPs to model}

\item{n_cores}{Number of cores to use}
}
\value{
Numeric matrix (haplotypes x frequency,snps,scans)
}
\description{
Get haplotypes with MCMC (linear complexity, multithreaded)
Jan 5, 2015
}
\author{
tcharlon
}

