% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.gds.class}
\alias{is_snp_first_dim.gds.class}
\title{test a snp gds file if the genotypes first dimension is snp}
\usage{
\method{is_snp_first_dim}{gds.class}(obj, ...)
}
\arguments{
\item{obj}{an object containing genotypes}

\item{...}{additional arguments}
}
\description{
test a snp gds file if the genotypes first dimension is snp
}

