% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.GenotypeData}
\alias{is_snp_first_dim.GenotypeData}
\title{test in a (gds) GenotypeData file if the genotypes first dimension is snp}
\usage{
\method{is_snp_first_dim}{GenotypeData}(obj, ...)
}
\arguments{
\item{obj}{an object containing genotypes}

\item{...}{additional arguments}
}
\description{
test in a (gds) GenotypeData file if the genotypes first dimension is snp
}

