% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.MatrixGenotypeReader}
\alias{is_snp_first_dim.MatrixGenotypeReader}
\title{always TRUE}
\usage{
\method{is_snp_first_dim}{MatrixGenotypeReader}(obj, ...)
}
\arguments{
\item{obj}{an object containing genotypes}

\item{...}{additional arguments}
}
\description{
always TRUE
}

