% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{load_gds_as_genotype_data}
\alias{load_gds_as_genotype_data}
\title{load  GDS file as GWASTools::GenotypeData object}
\usage{
load_gds_as_genotype_data(gds_file, read_snp_annot = TRUE,
  read_scan_annot = TRUE)
}
\arguments{
\item{gds_file}{the gds_file to load}

\item{read_snp_annot}{whether to read embedded SnpAnnotationDataFrame}

\item{read_scan_annot}{whether to read embedded ScanAnnotationDataFrame}
}
\value{
a GenotypeData object linked to a GdsGenotypeReader
}
\description{
If the the SnpAnnotationDataFrame and the ScanAnnotationDataFrame
were saved (cf save_genotype_data_as_gds)) they will be restored.
}
\seealso{
Other gds: \code{\link{save_genotype_data_as_gds}};
  \code{\link{write_genotypes_to_gds}};
  \code{\link{write_snp_gds}}
}

