% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc_pca.R
\name{merge_dfs}
\alias{merge_dfs}
\title{Merge data frames on rownames}
\usage{
merge_dfs(l_df, ...)
}
\arguments{
\item{l_df}{List of data frames}

\item{...}{Passed to merge}
}
\value{
Data frame
}
\description{
Merge data frames on rownames
}
\author{
tcharlon
}

