% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/5_methods_genotype_data_subset.R
\docType{methods}
\name{nscan,GenotypeDataSubset-method}
\alias{nscan,GenotypeDataSubset-method}
\title{GenotypeDataSubset wrapper}
\usage{
\S4method{nscan}{GenotypeDataSubset}(object)
}
\arguments{
\item{object}{Object, possibly derived from or containing
    \code{\link{NcdfReader-class}},
    \code{\link{GdsReader-class}},
    \code{\link{ScanAnnotationDataFrame-class}},
    \code{\link{SnpAnnotationDataFrame-class}},
    \code{\link{ScanAnnotationSQLite-class}}, or
    \code{\link{SnpAnnotationSQLite-class}}.}
}
\description{
GenotypeDataSubset wrapper
}

