% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{\%//\%}
\alias{\%//\%}
\title{an imitation of the perl // operator}
\usage{
a \%//\% b
}
\arguments{
\item{a}{any object}

\item{b}{any object}
}
\value{
a unless NULL, in which case b
}
\description{
like "NULL OR"
}
\examples{
\dontrun{
res <- arg1 \%//\% 1 # default value if arg1 is not defined
}
}

