% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_fortify}
\alias{pca_fortify}
\title{Fortify the original dataset used for the
pca with the results of the PCA.}
\usage{
pca_fortify(qb_pca, active_obs = TRUE, sup_obs = TRUE, active_vars = TRUE,
  sup_vars = TRUE, include_pc_variance = TRUE)
}
\arguments{
\item{qb_pca}{qb_pca object}

\item{active_obs}{Character identifiers of active observations to use. If TRUE no
subset. If FALSE none are used.
  Default: TRUE}

\item{sup_obs}{Character identifiers of suppl. observations to use. If TRUE no
subset. If FALSE none are used.
  Default: FALSE}

\item{active_vars}{Character identifiers of active variables to use. If TRUE no
subset. If FALSE none are used.
  Default: FALSE}

\item{sup_vars}{Character identifiers of suppl. variables to use. If TRUE no
subset. If FALSE none are used.
  Default: FALSE}

\item{include_pc_variance}{Logical, includes the explained variance in the data frame}
}
\value{
data.frame of original data with
additional columns containing the PCs.
The explained variance, variables, and individuals are rbinded in the
same data.frame
}
\description{
Fortify the original dataset used for the
pca with the results of the PCA.
}
\author{
adicara
}

