% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select.R
\name{peak_selection}
\alias{peak_selection}
\title{peak_selection}
\usage{
peak_selection(df_vars, chromosomes, n_cores = 2)
}
\arguments{
\item{df_vars}{Data frame of variable contributions with columns
PCA_VARNAME and PC{axes}}

\item{chromosomes}{Chromosomes of df_vars rows}

\item{n_cores}{Number of cores to use}
}
\value{
List of snpIDs selected, each slot for each principal component
}
\description{
peak_selection
}

