% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{read_snp_gds_alleles}
\alias{read_snp_gds_alleles}
\title{read and process the alleles stored in a snp gds file}
\usage{
read_snp_gds_alleles(gds)
}
\arguments{
\item{gds}{a gds object}
}
\value{
a character vector of alleles in the "A/B" format
     and NA for missing alleles, or NULL if not available
}
\description{
The "" empty alleles are decoded as NA
}

