% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{reorder_pcafort}
\alias{reorder_pcafort}
\title{Reorder dataframe with PCA obs (nas first)}
\usage{
reorder_pcafort(df_all_obs, nas_first, groups)
}
\arguments{
\item{df_all_obs}{sub_df of qbpcafort containing obs}

\item{nas_first}{Logical, rearranges data frame to put NAs in groups first}

\item{groups}{Column index or name of qb_pca$data for the grouping of
observations.
  Default: NULL}
}
\value{
df
}
\description{
Reorder dataframe with PCA obs (nas first)
}
\author{
tcharlon
}

