% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{request_snpgds_file}
\alias{request_snpgds_file}
\title{make a snpgds file if needed}
\usage{
request_snpgds_file(gdata, snps_idx = NULL, scans_idx = NULL)
}
\arguments{
\item{gdata}{the genotype data object to save}

\item{snps_idx}{indices of snps to use}

\item{scans_idx}{indices of scans to use}
}
\value{
a list(snpgds, snps_idx, scans_idx, new_file, snp_ids, scan_ids)
}
\description{
most SNPRelate functions use a snpgds file.
GenotypeData using a GdsGenotypeReader have already
a snpgds file that can be used, but others don't.
}
\details{
This function takes care of the details, and return a list
of (snpgds to use, snps_idx = the snp indices
to use on this file that are equivalent to the
input snps_idx, scans_idx= idem, new_file=whether the snpgds
is new, and ergo should be deleted, snp_ids: the snps ids to use
on the snpgds to use, scan_ids: the scan ids to use on the new file)

do not forget to delete the newly created file if any !!!!!
}
\author{
karl
}

