% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc_pca.R
\name{sample_impute}
\alias{sample_impute}
\title{Impute nas by sampling existing values}
\usage{
sample_impute(data)
}
\arguments{
\item{data}{matrix with nas}
}
\value{
matrix with imputed values
}
\description{
Jun 17, 2014
}
\author{
tcharlon
}

