% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{save_genotype_data_as_gds}
\alias{save_genotype_data_as_gds}
\title{save a GWASTools::GenotypeData object a a GDS file}
\usage{
save_genotype_data_as_gds(gdata, filename, save_annotations = TRUE,
  compress = "ZIP.max", check = TRUE, quiet = FALSE, chunk_size = 1000,
  ...)
}
\arguments{
\item{gdata}{the GenotypeData object to save}

\item{filename}{the gds output filename}

\item{save_annotations}{whether to include in the gds the serialization
of the SnpAnnotationDataFrame and
ScanAnnotationDataFrame if applicable}

\item{compress}{the compression format for the annotation in the
GDS file, one of "", "ZIP", "ZIP.fast", "ZIP.default", or "ZIP.max"}

\item{check}{whether to check the results (paranoid check)}

\item{quiet}{whether to be quiet}

\item{chunk_size}{the desired chunk size}

\item{...}{additional arguments for
                           \code{\link{fetch_genotypes_by_chunk_iterator}}}
}
\description{
WARNING: All NA in character vector or columns are converted into empty
strings ""
}
\seealso{
Other gds: \code{\link{load_gds_as_genotype_data}};
  \code{\link{write_genotypes_to_gds}};
  \code{\link{write_snp_gds}}
}

