% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{save_genotype_data_as_plink}
\alias{save_genotype_data_as_plink}
\title{save a GWASTools::GenotypeData object as binary PLINK files}
\usage{
save_genotype_data_as_plink(gdata, basename, dir = getwd(), snps_idx = NULL,
  scans_idx = NULL, ...)
}
\arguments{
\item{gdata}{the genotype data object to save}

\item{basename}{the basename/path of the files to save}

\item{dir}{the destination dir}

\item{snps_idx}{indices of snps to save}

\item{scans_idx}{indices of scans to save}

\item{...}{additional arguments for snpgdsGDS2BED}
}
\description{
Will create 3 files: .bed, .fam and .bim
}

