% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data2.R
\name{save_hgdp_as_gds}
\alias{save_hgdp_as_gds}
\title{save_hgdp_as_gds}
\usage{
save_hgdp_as_gds(paths = file.path(system.file("extdata", package =
  "snpclust"), paste0("hgdp.", c("zip", "txt", "gds"))),
  zippaths = paste0("hgdp/", c("HGDP_FinalReport_Forward.txt",
  "HGDP_Map.txt")))
}
\arguments{
\item{paths}{Paths of the zip and txt files}

\item{zippaths}{Paths of the genotype and snp files in the zip}
}
\value{
Path of the saved gds file
}
\description{
save_hgdp_as_gds
}

