% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{scale_pca_vars}
\alias{scale_pca_vars}
\title{Scale vars in qbpca to display them}
\usage{
scale_pca_vars(df_all_vars, names_axe, scale_sdev = FALSE, scale = 1)
}
\arguments{
\item{df_all_vars}{df with all variables (subset of qbpcafort)}

\item{names_axe}{names of PCA axes}

\item{scale_sdev}{Should the values be scaled by the std. dev. of PCs ?}

\item{scale}{scale of the values}
}
\value{
df similar to input one but with scaled vars
}
\description{
Scale vars in qbpca to display them
}
\author{
adicara
}

