% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_temp_dir}
\alias{setup_temp_dir}
\title{create a temp directory, that will automatically deleted}
\usage{
setup_temp_dir(chdir = TRUE, ...)
}
\arguments{
\item{chdir}{whether to set the current directory to the new temp dir}

\item{...}{args forwarded to \code{\link{tempfile}}}
}
\value{
the created directory path invisibly
}
\description{
when the calling function exits, and can set the working directory
temporarily into this new directory. The previous current directory will
be restored when the calling function exits.
}

