% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{snpclust-package}
\alias{snpclust-package}
\title{Single nucleotide polymorphism unsupervised clustering}
\description{
The R package snpclust performs unsupervised feature selection and summarization based on Principal Component Analysis, Gaussian Mixture Models, and Markov Chain Monte Carlo.
SNPClust is developed for the reclassification of Systemic Autoimmune Diseases based on genetic markers instead of clinical criteria for the European-funded project PRECISESADS (http://precisesads.eu).
The snpclust function is the main function and can be called on a snpgds file path from the R package SNPRelate.
Results can be displayed with several plots. A demo is available in the *SNPClust applied to Europeans* HTML vignette.
}
\keyword{package}

