% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpclust.R
\name{snpclust}
\alias{snpclust}
\title{snpclust}
\usage{
snpclust(
  gds,
  bed_paths,
  n_axes = 100,
  n_cores = 1,
  only_pca = FALSE,
  keep_tmp = FALSE,
  ...
)
}
\arguments{
\item{gds}{Path of Genomic Data Structure file}

\item{bed_paths}{Paths of PLINK binary files.
If missing, files are generated with the gds file.}

\item{n_axes}{Number of principal components to consider}

\item{n_cores}{Number of cores to use}

\item{only_pca}{Logical, compute only PCA and exit}

\item{keep_tmp}{Logical, keep temporary objects}

\item{...}{Passed to snprelate_qc}
}
\value{
List of slots:
          gdata: GenotypeData object of the quality controlled gds,
          qc: details about the quality control applied,
          pca: PCA applied to the quality controlled gds,
          peaks: list of SNP IDs selected in each principal component,
          features: matrix of selected SNPs and estimated haplotypes,
          features_pca: PCA applied to the features.
}
\description{
SNPClust performs PCA-based SNP selection and haplotype summarization of
nearby correlated SNPS in linear complexity using the SHAPEIT software.
Features are then weighted by pricipal component rank and contributions.
}
