% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snpclust.R
\name{snpclust_hgdp}
\alias{snpclust_hgdp}
\title{snpclust_hgdp}
\usage{
snpclust_hgdp(dir, n_cores = 2)
}
\arguments{
\item{dir}{Directory containing the HGDP files}

\item{n_cores}{Number of cores to use}
}
\value{
snpclust object
}
\description{
snpclust_hgdp
}

