% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc_pca.R
\name{snprelate_ld_select}
\alias{snprelate_ld_select}
\title{wrapper for snpgdsLDpruning to select Tag SNPs}
\usage{
snprelate_ld_select(gdata, window_length = 500L, min_r2, window_size = NA,
  snps_idx = NULL, scans_idx = NULL, remove.monosnp = FALSE,
  autosome.only = FALSE, method = "r", threads = 1, quiet = FALSE, ...)
}
\arguments{
\item{gdata}{a GenotypeData object}

\item{window_length}{max length in kb of the window}

\item{min_r2}{minimum r2 value to report}

\item{window_size}{max number of SNPs in LD window}

\item{snps_idx}{indices of snps to use}

\item{scans_idx}{indices of scans to use}

\item{remove.monosnp}{if TRUE, remove monomorphic SNPs}

\item{autosome.only}{if \code{TRUE}, use autosomal SNPs only; if it is a
        numeric or character value, keep SNPs according to the specified
        chromosome}

\item{method}{"composite", "r", "dprime", "corr", see details}

\item{threads}{the number of threads to use. currently ignored}

\item{quiet}{whether to be quiet}

\item{...}{forwarded to \code{\link[SNPRelate]{snpgdsLDpruning}}}
}
\value{
a list of SNP IDs stratified by chromosomes.
}
\description{
The tagged snp set is (by sliding window) representative and
strongly not redundant
}
\details{
will use the "r" method
#' N.B: it is not yet parallelized
}
\author{
karl
}

