% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprelate_qc_pca.R
\name{snprelate_pca}
\alias{snprelate_pca}
\title{snprelate_pca}
\usage{
snprelate_pca(gdata, n_axes = 32, n_cores = 1)
}
\arguments{
\item{gdata}{Genotype data object}

\item{n_axes}{Number of axes}

\item{n_cores}{Number of cores}
}
\value{
Data frame with variance, observations, and variables
}
\description{
Principal component analysis using SNPRelate
}
