% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc_pca.R
\name{snprelate_qc}
\alias{snprelate_qc}
\title{snprelate_qc}
\usage{
snprelate_qc(gdata, sample_nas = 0.03, snp_nas = 0.01, maf = 0.05,
  tsnp = 0.8, ibs = 1, n_cores = 2)
}
\arguments{
\item{gdata}{GenotypeData object}

\item{sample_nas}{Missing values ratio threshold for samples}

\item{snp_nas}{Missing values ratio threshold for SNPs}

\item{maf}{MAF ratio threshold}

\item{tsnp}{LD threshold}

\item{ibs}{Identity by state threshold}

\item{n_cores}{Number of cores to use (default detected)}
}
\value{
list of gdata and QC data frame
}
\description{
Oct 10, 2014
}
\author{
tcharlon
}

