% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{write_genotypes_to_gds}
\alias{write_genotypes_to_gds}
\title{write genotypes to a GDS file/node}
\usage{
write_genotypes_to_gds(gds, geno_iterator, node_name = "genotype",
  quiet = FALSE)
}
\arguments{
\item{gds}{a gds.class}

\item{geno_iterator}{a fetch_genotypes_by_chunk_iterator iterator}

\item{node_name}{the node of the name to create}

\item{quiet}{whether to be quiet}
}
\value{
the created gdsn node
}
\description{
write genotypes to a GDS file/node
}
\seealso{
Other gds: \code{\link{load_gds_as_genotype_data}};
  \code{\link{save_genotype_data_as_gds}};
  \code{\link{write_snp_gds}}
}

