% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geno_data.R
\name{write_snp_gds}
\alias{write_snp_gds}
\title{write a snp GDS file as used by SNPRelate}
\usage{
write_snp_gds(filename, geno_iterator, sample_ids, snp_infos,
  compress = "ZIP.max", quiet = FALSE, ...)
}
\arguments{
\item{filename}{the filename of GDS file to create}

\item{geno_iterator}{a fetch_genotypes_by_chunk_iterator iterator}

\item{sample_ids}{the sample ids, must be unique}

\item{snp_infos}{a data frame with columns snpID, chromosome,
position, alleles}

\item{compress}{the compression method for the variables except genotype}

\item{quiet}{whether to be quiet}

\item{...}{options forwarded to \code{\link{write_genotypes_to_gds}}}
}
\description{
This is a new implementation of snpgdsCreateGeno optimized for memory
}
\details{
N.B: the genotypes are stored in sample.order, i.e optimized to read
SNP by SNP, not sample by sample

The alleles must be specified as character "A/B". e.g. "C/T"
}
\seealso{
Other gds: \code{\link{load_gds_as_genotype_data}};
  \code{\link{save_genotype_data_as_gds}};
  \code{\link{write_genotypes_to_gds}}
}

