% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{get_pca_panels}
\alias{get_pca_panels}
\title{Get list of pc plots}
\usage{
get_pca_panels(pca, axes = 1:4, groups = NULL, max_vars = 0L, ...)
}
\arguments{
\item{pca}{qb_pca or pca object}

\item{axes}{pcs to combine}

\item{groups}{Column index or name of qb_pca$data for the grouping of
observations.
  Default: NULL}

\item{max_vars}{max. number of vars to display}

\item{...}{Passed to ggplot_pca}
}
\value{
list of pca ggplots
}
\description{
Jul 9, 2014
}
\author{
tcharlon
}

