% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{ggplot_manhat}
\alias{ggplot_manhat}
\title{ggplot_manhat}
\usage{
ggplot_manhat(pca, gdata, axes = paste0("PC", 1:10), byposition = TRUE,
  n_points = 2000, ncol = 5, ...)
}
\arguments{
\item{pca}{PCA data frame, requires variable contributions}

\item{gdata}{GenotypeData object with snpIDs matching df_vars column
PCA_VARNAME}

\item{axes}{Axes to plot}

\item{byposition}{Plot the SNPs by chromosome and poistion or by index}

\item{n_points}{Maximum number of largest values to plot, for each
principal component}

\item{ncol}{ggplot2::facet_wrap}

\item{...}{Passed to ggplot2::facet_wrap}
}
\value{
ggplot
}
\description{
Displays Manhattan-like plot for SNPs contributions to Principal Components
}

