% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplots.R
\name{plot_pca_pairs}
\alias{plot_pca_pairs}
\title{Plot PCA pairs}
\usage{
plot_pca_pairs(axes, ..., max_vars = 0L, ellipses = TRUE)
}
\arguments{
\item{axes}{pcs to combine}

\item{...}{Passed to get_pca_panels}

\item{max_vars}{max. number of vars to display}

\item{ellipses}{Should ellipses be plotted ? Not used if groups is NULL.
Default: FALSE}
}
\description{
Plot PCA pairs
}

