% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{plot_pca_pairs}
\alias{plot_pca_pairs}
\title{Plot PCA pairs}
\usage{
plot_pca_pairs(axes, ..., max_vars = 0L, ellipses = TRUE)
}
\arguments{
\item{axes}{Two principal component axes to plot.
Default: 1:2}

\item{...}{Passed to get_pca_panels}

\item{max_vars}{Number of variables to display}

\item{ellipses}{Should ellipses be plotted ? Not used if groups is NULL.
Default: FALSE}
}
\description{
Plots the pair combinations of Principal Components.
}

