% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snpclust.R
\name{snpclust}
\alias{snpclust}
\title{snpclust}
\usage{
snpclust(tar_paths, gds, pca_names = "", n_axes = 100, n_cores = 2)
}
\arguments{
\item{tar_paths}{Path(s) of tar.gz files containing PLINK binary files}

\item{gds}{Path of Genomic Data Structure file}

\item{pca_names}{Character vector for subsetting in get_functional_pca}

\item{n_axes}{Number of principal components to consider}

\item{n_cores}{Number of cores to use}
}
\value{
List of slots pca, qc, and features
}
\description{
Feb 10, 2015
}
\author{
tcharlon
}

