% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snpclust.R
\name{snpclust}
\alias{snpclust}
\title{snpclust}
\usage{
snpclust(tar_paths, gds, subsets = "", n_axes = 100, n_cores = 2, ...)
}
\arguments{
\item{tar_paths}{Path(s) of tar.gz files containing PLINK binary files.
If missing, files are generated with the gds file.}

\item{gds}{Path of Genomic Data Structure file}

\item{subsets}{Character vector for subsetting the dataset.
'' - no subset
'hla' - use only SNPs in HLA
'non_hla' - use only SNPs outside HLA
'no_controls' - exclude samples with value Control in
phenotype}

\item{n_axes}{Number of principal components to consider}

\item{n_cores}{Number of cores to use}

\item{...}{Passed to snprelate_qc}
}
\value{
List of slots:
          pca - PCA applied to the quality controlled dataset,
          qc - details about the quality control applied,
          gdata - GenotypeData object of the quality controlled dataset,
          peaks - list of SNP IDs selected in each principal component,
          features - matrix of SNPClust features before weighting,
          and features_pca - PCA applied to the weighted SNPClust features.
        If several subsets are provided, each slot will be a list with
        slots corresponding to each subsets.
}
\description{
Feb 10, 2015
}
\author{
tcharlon
}

