% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc_pca.R
\name{snprelate_pca}
\alias{snprelate_pca}
\title{snprelate_pca}
\usage{
snprelate_pca(gdata, n_axes = 32, n_cores = 2)
}
\arguments{
\item{gdata}{GenotypeData object}

\item{n_axes}{Number of PCs to fetch}

\item{n_cores}{Number of cores}
}
\value{
qb_pcafort
}
\description{
Oct 10, 2014
}
\author{
tcharlon
}

